PDF interpreter - cope with strangely constructed Pages trees
Bug #700953 "Error: /execstackoverflow in --.systemvar--"

The pre-interpretation scan of the Pages tree looking for loops in the
tree structure uses recursion; if a Pages tree is badly constructed it
is possible to recurse enough times to cause an overflow of the exec
stack.

We can cope with this by aborting the scan and continuing to interpret
and render the PDF file, because the regular processing of PDF file
doesn't use recursion.

Of course, this does mean that we lose the check for loops in the Pages
tree (at least, loops beyond the point where we overflow the exec stack)
but both conditions are rare, so I think its worth doing.

We do still raise a warning.