Tweak the interaction of file permissions and -I
The new file permissions code differentiates between 'path', 'path/' and
'path/*', whilst the old -I code did not.

This slightly modifies how the -I stuff drives the file permissions stuff so it
works (more) like the previous -I behaviour.

The remaining difference is that the new code always treats '*' as a wildcard,
where the old code would permit '*' as a 'real' character in a path.