pdfwrite - handle nested Patterns inside a transparency Group
Bug #701181 "Incorrect output after ghostscript processing of a PDF file"

The PDF file is constructed in a somewhat sub-optimal fashion. Each
background square 'grid' is a Form, each Form draws the content by
filling the area with a solid colour, and then drawing white lines
across it. The white lines (actually rectangles) are drawn using a
Pattern.

So far so good, however the Pattern simply executes a Form, that Form
then fills another rectangle, with another Pattern, its this final
pattern which does the drawing.

This is two levels of indirection more than required.

Normally this isn't a problem because we don't attempt to preserve
Forms from PDF files (and this is one example of why we don't). But....
If the file involves transparency Groups, then we must use a Form
XObject for the Group.

In this case the CTM of the first pattern is the CTM of the 'enclosing
context', ie the Form, but the CTM of the second pattern is also the
CTM of the 'enclosing context', but now that means the first pattern,
not the Group.

To cater for this we track the number of patterns defined since the last
Form definition. If its 0 (ie this is the first pattern since a Form
definition) then we just use the CTM. If its greatre than 0, however,
then we 'undo' the scaling applied by pdfwrite, in order to get back to
the CTM of the enclosing pattern.

Its possible that more extreme examples (patterns inside patterns
inside groups inside patterns inside groups) might still not be entirely
correct, but we have no examples of this (Patterns inside patterns are
relatively rare). If we find any we can revisit this.