Fix segfault and some cleanups related to DefaultQState and patterns
Fixed segfault on pgmraw device with tests_private/comparefiles/Bug691763.pdf

The issue was that the saved ctx->base_pgs (now renamed DefaultQState
to match ps code) was saved before the transparency device was
installed, so it had the wrong device in it for the transparency case
(wreaking havoc).

Added better support for "DefaultQState", so it will have the correct
device in it.  Added accessor functions.

Change to set the DefaultQState after transparency has been handled, so the
device will be correct.

Also changed a gsave/grestore to a q/Q to match gs implementation.