pdfwrite - don't permit fallback font substitution with PDF/A
Bug #701373 "PDFA fails to validate by Preflight"

The inptu is a PDF file which is missing two fonts, Helvetica and
Calibri. For both fonts we use /Helvetica as a substitute. For the
missing Helvetica font this is correct, but for the missing Calibri font
this leads to problem when creating PDF/A otuput files.

Firstly, we now have two different fonts, with differing /Widths which
are potentially attempting to reference the same glyphs from a single
font. Secondly, no matter what we do, the /Widths from the Calibri font
are not going to match the actual glyph metrics from Helvetica.

Since there is no possible way to deal with the second problem, because
either the metrics won't match the /Widths, or the text will be misplaced
if we use Widths that are not those from Calibri, there is no point in
trying to tackle the first problem.

So here we introduce a new special_op 'EventInfo' which passes a single
name to the device. The name is the name of the event which occurred. For
now this can only be 'SubstitutedFont'. The only device which handles
this event is pdfwrite, all other devices simply ignore it. The pdfwrite
device checks to see if it is creating a PDF/A otuptu file, and if it is
then it emits a warnign and aborts the production of PDF/A.