Transparency compositor - forward unhandled special_ops to target device
Noticed while doing earlier work. The pdf14 device only passed on a few
specific gxdso_ special_op types to the target device, the remainder
were sent to the default handler. This seems wrong and caused problems
for my new gxdso at the time. Instead of it going through the pdf14
device and ending up at pdfwrite (or the default handler for other
devices) it was always going to the default handler and never making
it to pdfwrite when rendering transparency.

The pdf14 device should instead handle any gxdso_ special_ops that vary
depending on rendering. For instance the JPEG passthrough should *not*
be passed to the target device, the pdf14 device needs to have the
uncompressed image data to render it to the compositor buffer.

This commit adds handlers for such gxdso_ special ops, and passes all
the remaining unhandled ones to the target device for processing.

For some reason this causes diffs in halftones, I have no idea why. but
they don't seem like problems.