PDF interpreter - ignore visibility of annotations for devices which support them
No bug for this, arose from a Stack Overflow question, but its a FAQ
as well.

It used to be the case that the PDF interpreter would (and should)
render files containing annotations with the Print and NoView flags
differently, depending on whether the destination was to be rendered
or sent to a high level device. This is controlled with the -dPrinted
flag to determine whether annotations should be treated as for a Print
device or a View device.

This doesn't make a lot of sense for pdfwrite though. We really want to
keep the Annotation 'as is' and let the viewer or printer of the output
PDF file deal with it.

In the past we've moved the annotation handling for pdfwrite and,
technically, for any device which directly supports annotations from
'drawannot' to 'preserveannot'. Now that we've done that, it makes some
sense to have the 'preserveannot' code no longer check the annotation's
visibility, but simply preserve it as you would expect. NOTE we do check
the visibility state of annotations that we don't know how to handle.
These will not be preserved as annotations, but will be drawn into the
page contents, so we don't want to render any which should be invisible.
Such annotations are still affected by -dPritned.