Capture args to feed to pdfwrite.
To facilitate debugging, we want a way to pickle the gs invocation
used into device output (for devices that want it).

We therefore amend the argument processing logic to pass arguments
to gs_lib_ctx, where they will are 'sanitized' and stored in a list.
This list can then be accessed by a device.

We are careful to sanitize the args so as not to reveal paths or
secret passwords/code etc. Essentially we never store any filenames
or string values stored (except for certain whitelisted -s options).

It is likely that we'll want to expand the whitelisted values in
future.