Document the new SAFER/file access controls
The documentation for the old SAFER remains in place (noted as deprecated) until
we actually remove it. Document accessing the old SAFER functionality
(-dOLDSAFER).

Generally, tidy up the HTML/formatting of the parameter and operator lists
in Use.htm and Language.htm - mainly making indenting even and consistent.

List .setpdfwrite as deprecated

Include setting the target of the first parameter of gsapi_new_instance() to
NULL before calling it, and fix the examples in API.htm