Deprecate .setpdfwrite
The .setpdfwrite operator nowadays does nothing more than increase the
VM threshold, which is relatively pointless so we're going to remove it.

In this commit; add a warning message to the .setpdfwrite operator so
that anyone using it will be warned. Remove usage from the various
shell scripts we provide.