pdfwrite/ps2write/pdfimage - record args used
Following on from commit 813d7e96c8733f85d59a62732171cf7cedfc087d this
adds functionality to the pdfwrite and pdfimage families of devices to
actually write the args that were used to generate the output file into
the final output file, as comments.

Output files are somewhat larger with this commit, but only very small
files are in any way significantly larger. We do have a number of such
files in our test suite which exhibit larger output ( > 5%increase).