ps2write - don't password encrypt PostScript output!
Noticed while testing the args commit; if we set -sOwnerPassword (to
create a password-protected PDF file), but select (e)ps2write as the
device, then the resulting PostScript file would have all the streams
(in a PDF sense) encrypted with PDF password encryption.

Unsurprisingly, this result in a completely unusable PostScript file.
Lets not permit that....