Supporting "GroupGState" and "GroupMat" for SMask
Need to save the gstate at the time that the GS_SMask is set, so
it needs to go into the client_data in the gstate.

What a mess!  I think I have this correct, but it's ugly.  Need to make
sure the gstate that we are saving doesn't have a client_data reference or
havoc ensues.  I think this is fine because in the context where it's
needed, the SMask should be nulled out.  If we ever put additional things
in client_data, may need to revisit this.

This commit has some visual progressions, some visual regressions, and
fixes some segfaults... I have made note of some of the visual regressions
and will look into them.