Fix X11 device ICC profile memory leaks
When the x11 devices are running in "buffered mode" (i.e. we render to a
pixmap, then blit the pixmap to the X11 Window), we create a memory device
which actually does the rendering.

The memory device takes a reference to the ICC profiles from the x11 device
and it is reference counted.

Firstly, we always incremented the reference count, ignoring the posibility
that we were resizing an existing memory device rather than creating a new one.

Secondly, when shutting down the x11 device, we ignored posibility that
buffering was in force, and failed to free the memory device - most of the
memory was hoovered up by the garbager, but the ICC profile data isn't in
gc memory, hence showing a leak.

So, fix the memory device ICC reference counting, and reconfigure the device
as non-buffering before we shut it down.