Handle drag'n'drop file name/paths
Because we now run by default with the file access permission active,
we need to do additional work for Windows drag and drop files, so they don't
throw an invalidaccess error (since we implement drag and drop by sending
characters to the gs console to do '(file) run').

The problem is, we cannot add the file to read list, send the characters to the
console, and remove the file from the read list because, although SendMessage()
blocks until the message is handles, WriteConsoleInput() does not block until
the console buffers are consumed (i.e. it is asynchronous). So, there is no
certainty when the final SendMessage() in WM_DROPFILES case is finished that
Ghostcript will actually have run the file.

So, we create a list of dropped file names, add them to the permit read list,
when the next WM_DROPFILES event happens, or a WM_DESTROY event, we drop
file names from the permit read list before, if necessary, adding the current
ones.