Hack to disable counting down font in pdfi_grestore().
This is a temporary fix, probably causes memory leaks.

The crash is probably because it leaves a dangling reference in some
circumstance.  Unfortunately the test sample was too complicated to
easily pin down the circumstances.   I am not sure what the code here is
trying to accomplish (Ken?).

Fixes crash on this sample (debugbin/gpdf):

-dPDFDEBUG -Z:vg -sDEVICE=ppmraw -dMaxBitmap=4000m -r300
~/work/samples/tests_private/comparefiles/Bug693541.pdf

See also: tests_private/comparefiles/Bug688807.pdf