Handle explicit '\' escaping in path matching
As well as having to cope with both '/' and '\' as directory separators on
Windows, it turns out we also need to cope with the backslash separator being
explicitly escaped ("\\") - the explicit is added by the Postcript path
templating code in gs_init.ps.

I don't want to change templating code because that would likely have
implications in a lot of other places.