Windows Installer: cope with COMPILE_INITS=0
We default to COMPILE_INITS=1, and that was the only case that the installer
handled - i.e. building the installer with COMPILE_INITS=0 would install an
executable that wouldn't work without extra action (-I.. or similar).

This changes the installer compiler invocation, and the installer creation
so it behaves as it does now with COMPILE_INITS=1, but with COMPILE_INITS=0 the
installer will write the path to Resource/Init into the GS_LIB registry
key.