PDF interpreter - more /Mask strangeness in Acrobat
Bug #701468 " More weirdness with images containing invalid /Mask values"

In order to render this file as per Acrobat, we need to ignore the /Mask
array. However, in order to render bug 689717 correctly we need to fix
the /Mask array.

Both cases are 1 BitsPerComponent DeviceGray images, the only difference
is the actual /Mask values. Bug 689717 has [255 255] while 701468 has
[243 255]. We can differentiate based on whether we have DeviceGray
1 BPC image, where the Mask values are not the same. In the case they
are different we use the existing technique of calculating the maximum
value and using that (for both, so there is no range). Where there is a
range (ie not a single colour) this is clearly unusable in a 2-colour
image, so ignore the /Mask altogether.