pdfwrite - fix corner case with Color Conversion
Bug #701456 "Ghostscript creates broken PDF file after color conversion"

The test file is, essentially weird. The image on page 3 is not a single
image, its a series of images, all of which are in a /Indexed /DeviceRGB
colour space.

In addition, these images are drawn on *top* of another set of images
(thereby completely obscuring them). These images are also drawn in
/Indexed /DeviceRGB. However, these images are in fact monochrome, the
Indexed space consists of 2 colours. This means that (highly unusually)
we have a DeviceRGB image with 1 Bit Per Component.

This caused problems for the pdfwrite compression filter chooser because
it chose (and configured) a compression scheme suitable for 1 BPC data
and then actually output 8 BPC data. This led to the compression filter
writing too little data.

Fix this here by setting the BPC to 8 if we are doing colour conversion,
and have the colour conversion setup code use the original value in the
image enumerator (the conversion code *does* need to know the input is
1 BPC).