Fix font file finding with -P
In trying to reproduce the issue that inspired commit dea69cd04964, I noticed
that font file searching wasn't working correctly with -P specified on the
command line.

Previously, it could just end up with us not finding a font we should have, but
with dea69cd04964 it would result in a typecheck error.

This ensures the stack is always how it should be after the findlibfile
call