Support 'BC' and 'S' params in SMask
General cleanup of the code, to more closely match the spec.  The PS
code does all kinds of nonsense translating things that is mostly not needed.
Also made a note of the page in the PDF doc because it is a PITA to
find the spec for this otherwise (Index fail..)

Also fixed declaration of some name/string operators in pdf_int.c
(didn't hurt anything, but confusing).