Fix clearing of the drag and drop file list
Because we have to store the list of file names from drag and drop events, in
order to correctly add and remove them from the permit file read list, we also
want to clear the lists before we destroy the "text" window object.

The problem is, this happens after we shutdown Ghostscript and unload the dll.

This moves that clearing of the list into function that we call before we
shut down.