Better transparency support for stroke/fill operators
Changes for setup_trans, teardown_trans and BM, though the BM stuff
isn't implemented yet.

Also some refactoring of the stroke/fill reoutines to reduce duplicate code.

This shows a lot of progressions, and also a few suspicious things.

I noticed that a lot of my BBox's are off by like 0.5.  Need to figure
out what is going on there.