Implement TR, TR2, BG, BG2, UCR and UCR2
In order to use the '2' variants, we need to store the mapping in
force at the start of the page. So at the start of every page we take
a copy of the proc and map values, so we can use those if needed.

To implement the functions, we either set the gs_identity_transfer,
saved function (Default) or gs_mapped_transfer function. If we are
using the Default, then we simply copy the map values across. If we
have been given a function, then we evaluate the function for inputs
ranging form 0 to 1, in as many steps as a re required to satisfy
transfer_map_size. We store each result in the map.

This results in 2 files now causing seg faults in the PDf14 device at
high resolution (clist). However I think this is simply due to memory
changes, I don't believe the code itself causes the problems.