Change filladjust to 0.3
It is being set twice.  First time it was correct at 0.3, second
time it was being overridden to 0.  Not sure why the second one is
even there, but it appears it should be 0.3.

This is going to generate a gazillion diffs, but should be a really good
change. Some of my test samples that looked correct, but didn't bmpcmp
with the gs version, now bmpcmp exactly!