Bug 701563: pdfwrite annotation pdfmark /Border key
Despite the spec saying the values for the /Border key are in user space
coordinates, Distiller appears to pass the values through to the output PDF
untouched (no rescaling - unlike the values for /Rect, also in user space,
according to the spec, which it does scale/translate). Ghostscript/pdfwrite
did scale the /Border values.

This just changes pdfwrite to write the values out unchanged.