Bug 701561: 'reduce' paths before adding to permit lists
Before attempting to open files, we pre-process the requested file name to
remove surplus/unnecessary elements: i.e. './././file' would be reduced to just
'./file', or '../dir/../dir/../dir/file' would be reduced to '../dir/file'.

The 'reduced' path is what we try to open, hence it is also what we check
against the file permissions list before we allow files to be accessed.

That being so, we should also 'reduce' paths as we add them to the permissions
lists - thus the permissions list creation and checking are consistent.