Bug 701568: Fix gdevpx.c RLE stream handling.
The current code in pclxl_write_image_data_RLE passes
lines of data to the RLE compression routine. It tells
each invocation of that routine that this is the "last"
block of data, when clearly it is not.

Accordingly, the compression routine inserts the "EOD" byte
into the stream, and returns EOFC.

Independently of the return value used, having multiple EOD
bytes in the data is clearly wrong. Update the caller to only
pass "last" in for the last block.

The code still returns EOFC at the end of the data, so update
this final call to accept (indeed, expect) that return value
there.