Documentation - add a note about SAFER for PDF/X ICC profiles
Bug #701553 "Examples of PDF/X and PDF/A conversion should be updated for Ghostscript 9.28"

ICC profiles required for PDF/X and PDF/A output need to be readable
in order to work when -dSAFER is true (the new default). This wasn't
clear from the documentation, so add a note about it.

As noted in the bug report, it is possible to embed the profile in the
PostScript code, but as Chris says this is probably beyond reasonable
expectations of the average user; they simply won't change the example
and will then complain that it isn't correct. Users capable of making
this work almost certainly have enough knowledge to realise they can
do this instead of making the profile readable.