Fix file permissions interaction with '@' command line
If an options file (prefixed '@' on the command line) followed something that
required initialising the Postscript interpreter (such as a '-c' option),
opening the file would potentially fail because file permissions had been
activated by the interpreter.

Add and remove the file from the permit reading list before attempting to open
it.