Add pdf_optcontent.c and .h, move code around
Move ocg stuff from pdf_page to pdf_optcontent

Now pdf_page.[ch] is empty...  I think pdf_optcontent is a better place
for this, but not deleting pdf_page for now in case they turn out to
be useful for something?

Have image debug print out the object number

This is a bunch of code moving and renaming, no functional differences.