Implement code to handle OC in content sequences
Implement opcodes BMC, BDC, EMC
This is an incomplete implementation, just enough to match what the gs
code does, which is just supporting the OC functionality.
Basically this implements "OFFlevels".

The gs implementation uses a dictionary to keep track of which levels are
turned off.  I used an array for this (with ability to expand it if
needed).  Couldn't use pdf_dict because it has no 'undef' capability, and
actually the current implementation is really just an array anyway.

Add support for OFFlevels to these marking operators:
Fill, stroke, image, shading
Put in a TODO comment for TJ (for ken)

Also, minor refactor to fill/eofill.