PDF interpreter - handle ObjStm containing more objects than declared
There's no bug number for this one, I can't easily create a reduced file
without inadvertently fixing the problem, and I don't want to put the
customer file into Bugzilla.

The problem is that the file contains an ObjStm (a compressed stream
containing a sequence of PDF objects) which contains more objects than
the stream declares. Not only that, but the extra object is neither the
first nor the last object, so its not trivial to ignore.

Most consumers handle ObjStm constructs by using the 'catalog' at the
head of the stream to find the offset of each object in the stream and
then reposition to that object and read a single PDF object (note that
objects in an ObjStm do not begin with 'obj' and do not end with
'endobj') which is IMO putting a lot of faith in the consumer's
ability to detect a single object.

For performance reasons Ghostscript's PDF interpreter reads *all* the
objects in the ObjStm immediately and stores them, thus trading memory
use for performance (we only need to read and decompress the objects
once). This approach is totally defeated when there are more objects
than expected, because we don't know which object relates to which
object number (no 'x y obj' in an ObjStm).

In this commit we spot the discrepancy; if there are too few objects we
exit immediately with an error (technically we could do something about
this too, provided the missing object is never used). If there are extra
objects then we move to a fallback.

We discard all the objects we've read, then use the offsets in the
'catalog' to calculate the position and length of each object in the
ObjStm. We position the file pointer to the start of the object and
use a SubFileDecode filter to limit the number of bytes read to the
difference between the offsets. We then use the existing code to read
all the objects in that stream. If we get more than one then we discard
all but the first.

Repeat the above until we reach the last object in the catalog; that has
no known length so just read from there to the end of the stream.

The complications are all around getting the objects on the stack in
the order expected by the code which follows, and defines the actual
objects.