Bug 578865: Push transparency group for non idempotent shadings.
Shadings that write the same pixels more than once can be
rendered incorrectly for blendmodes that aren't idempotent
(or for non solid opacities). In such cases, push a transparency
group so that the shading can be written just once, and
then safely blended.

Note that this doesn't capture shadings that are rendered
via zshfill.