Fix "permit-file-xxxx" handling.
Ray spotted that gs_add_explicit_control_path was adding 17
to arg before using it. 17 happens to be the right amount
to add for "--permit-file-read=", but the wrong amount for
write/control/all.

Update the code to call it with the correct arg pointer
to start with.

Also, update a couple of routines to cope with being called
with NULL strings.

Also use enum values in switch rather than 0, 1, 2.