Remove spurious call to gp_validate_path_len()...
...in gx_device_open_output_file()

There is no need to validate the path in gx_device_open_output_file() as it will
be done at a lower level (gp_fopen) for paths that require it (we only validate
*after* the Postscript style %device% paths have been expanded into 'real'
paths).

This also interacts badly with the revised handling of formatted string
file names, hence remove it.