Knockout groups should knockout the group alpha (and shape) too.
When marking within a knockout group, for each pixel we mark, we
throw away the current colorant values in the destination, and
base the calculations of the new colorant values on the stored
backdrop values.

We were failing to do this for the group alpha and shape
calculations. This could lead to the group alpha getting out of
sync with the actual colorant values. We were seeing this with
shadings (particularly gradient shadings) where the shadings
paint some pixels more than once.

This becomes much more obvious with the forthcoming commit to
use knockout groups for shadings with either transparency or
non-idempotent blend modes.