Bug 701634: Correctly handle file name of "\0"
Attempting to open a file with a filename string starting with a NULL
character should be treated the same as if the string was zero length.

Not doing so created a broken file object leading to several operations
resulting in a segfault.

Also, add cleanup in the event of such an error, freeing memory allocated in
preparing the gs stream object.