jbig2dec: Invite user to define MEMENTO in CFLAGS.
Previously we invited the user to define MEMENTO in jbig2_priv.h.
The problem with doing that is that the flag is enabled for all .c
modules, except memento.c. This implies that while memento's wrapper
functions are called, the versions being called are the ones that
have memento features disabled.

Setting CFLAGS=-DMEMENTO while configuring/building is the better
option as CFLAGS will affect all .c modules.