pdfwrite - don't honour /Producer key in DOCINFO pdfmark
Bug #701639 "pdfwrite should not honour the /Producer key in DOCINFO pdfmark"

The Producer in the document information dictionary (and XML Metadata)
is defined quite clearly as the application which produced the PDF file
from its native format.

Ghostscript uses the Producer key (if present in the pdfmark) to set
the Producer, but it should not do so. Adobe Acrobat Distiller (at
least the end user version) does not permit this value to be altered.

On reflection we can see why; in the case of a problem with a PDF file
its important to know which application created it, and we should not
let PostScript obscure that (we already do not pass the Producer
information on when the input is PDF).

The code did already attempt to overwrite any Producer string which
contained 'Distiller' (I have no idea why) and that code contained an
error which could lead to an invalid Document Information dictionary
being created.

This commit removes the ability for DOCINFO pdfmarks to alter the
Producer, which obviously fixes the bug.

However..... it is actually important for our commercial customers to
be able to set this value. A problem with a PDF file created by one of
our customer's products should be reported to that customer, not us, as
we will not be able to investigate the problem while our customer
should be able to. At the very least our customer will know how to
retrieve the configuration of Ghostscript being used.

So permit the commercial version of Ghostscript to set the /Producer
from a pdfmark.