Fix memory corruption setting a halftone
When setting a new halftone in the graphics state, we try to re-use the data
from the existing device halftone.

The problem is that the device halftone can have higher component indices than
there are components in the new halftone we are creating. In this case, we can
end up writing off the end of the components array for the new halftone
structure.

Simply check that the new halftone has enough components before doing the
duplication.