Interim commit
Working on clipping text rendering modes, reformatting the code in
preparation for executing the 'draw' twice, which will be required for
clipping modes.

Noticed that the pdfwrite case (ctx->preserve_tr_mode) was swapping to
the fill colour before drawing the text (surprisingly this is required)
but wasn't swapping back again afterwards. This was causing some tex
to be stored in the wrong colour in the PDF file.

This shows a number of progressions, but often not clearly as the font
is not yet imlemented.