Another interim commit
This, I believe, now handles text rendering modes involving clipping
correctly. There are a number of progressions, mainly due to not trying
to clip text when we don't actually handle the font yet, but there are
some genuine progressions.

The code needs tidying up but with a staff meeting coming up I likely
won't get the time to do that for a while, and I don;t want to lose
this work.