PDF interpreter - Improve PreserveAnnots
PreserveAnnots currenlty works in conjunction with ShowAnnotTypes, but
We should really have a separate control for this.

This commit adds a new /PreserveAnnotTypses (must be set from PostScript
in the same way as ShowAnnotTypes). PreserveAnnots and PreserveAnnotTypes
only work with a device which can preserve Annotations (currently this
is the pdfwrite device only).

If the device supports Annotation preservation, and PreserveAnnots is
true (the default0 then for each annotation we consult the Types array,
if the annotation is to be preserved then we pass it to the device. If
not, then we pass the annotation to the 'draw' routines which will
render teh annotation instead.

Using the two separate controls allows us to selectively control, by
type, whther annotations are preserved, rendered, or simply ignored
altogether.

Also fix an omission, and some typos, in the documentation for
ShowAnnots in Use.htm