pdfwrite - Use CA and ca for text with transparency
Bug #701631 "pdfwrite: mishandling of alphas with Text render mode 2"

The problem is in both the PDF interpreter and pdfwrite. In the past
the PDF interpreter has never been able to create a fill+stroke
primitive, because PostScript doesn't support that. However we added
the ability for pdfwrite to maintain text rendering modes, for reasons
of quality, which breaks that assumption.

Although we added the stroke and fill constant alpha to the C graphics
state when we moved the PDF graphics state into the C graphics state,
nothing actually ever used them. Instead the PDF interpreter sets a
single 'opacity' alpha channel, based on whether the current operation
is a stroke or a fill. Because we could not emit a combined operation
this worked as we would set the opacity separately for each operation.

Of course for pdfwrite this doesn't work, and looking to the future
when the fill+stroke primitive is available in the graphics library,
we will need to do away with the opacity.alpha and properly use the
fill and stroke alphas stored in the graphics state. For now, however,
we need to continue to use the opacity.alpha, but when we are handling
text (in pdfwrite) actually use the values of stroke and fill alpha
instead of opacity.alpha.

So add a 'for_text' boolean to all the routines whihch need to set or
update the alpha, and when writing the ExtGState, if we are handling
text then use CA and ca instead of opacity.alpha.

This has also mean altering the XPS interpreter so that it sets the
stroke and fill alphas as well as opacity.alpha, otherwise pdfwrite
doesn't know which value to use for text operations.