pdfwrite - preserve rendering intent
bug #701487 "Renedring intent not preserved by pdfwrite"

The pdfwrite device was not previously handling rendering intent changes
this commit adds a rendering_intent member to the pdfwrite device
structure, initialises it to 1 (absolute colorimetric) and then tracks
changes to the graphics state, emitting /RI entries in ExtGSatet
dictionaries as required.

This shows a number of small progressions in the test suite, and marked
progressions with icc_v4_profile.pdf and icc_rendering_intent.pdf.