Bug 701657: Allocate "memory" device in immovable gc memory
Devices are supposed to be allocated in memory that cannot relocate (i.e.
non-gc memory or an immovable allocation). The memory device accessible from
Postscript (makeimagedevice/makewordimagedevice) was being created in regular,
relocatable gc memory.

This caused an invalid memory access when cleaning up subclassed devices, as the
one device in the chain may have moved. So switch to using an immovable
allocation.