PDF interpreter - fix /None case of Line Ending style
Bug #701622 "pdfwrite device fails with annotation appearance"

This is, in fact, nothing at all to do with pdfwrite. The problem is
that, when generating an Appearance stream for a /Line Annotation which
is defined without an Appearance stream, the /None line ending style
was not popping the copy of the annotation dictionary.

This led to an error in the annotation. If a previous annotation used
transparency, then this would cause the compositor to get confused,
because we would start the transparency in the Line annotation, but
the error meant we would not close the transparency group.